Add GHC 9.10 support
Bumping the dependency bounds for `base` and `template-haskell` is
sufficient.  There are no changes to the code.

I noticed that the README was out-of-date, so this commit updates the
README as well as CHANGELOG.  Please note that CHANGELOG update is in
a new "Unreleased" section that needs to be updated the with the
version and date upon release.